+++ app/lib/models/host.dart
@@ -23,6 +23,7 @@ class Host {
   bool autoRecord;
   String? jumpHostId;
   bool shellIntegration;
+  bool agentForwarding;
   SftpMode sftpMode;
   String? sftpServerCommand;
 
@@ -41,6 +42,7 @@ class Host {
     this.autoRecord = false,
     this.jumpHostId,
     this.shellIntegration = true,
+    this.agentForwarding = false,
     this.sftpMode = SftpMode.normal,
     this.sftpServerCommand,
   })  : id = id ?? const Uuid().v4(),
@@ -64,6 +66,7 @@ class Host {
         'autoRecord': autoRecord,
         'jumpHostId': jumpHostId,
         'shellIntegration': shellIntegration,
+        'agentForwarding': agentForwarding,
         'sftpMode': sftpMode.name,
         'sftpServerCommand': sftpServerCommand,
       };
@@ -114,6 +117,7 @@ class Host {
       autoRecord: (json['autoRecord'] as bool?) ?? false,
       jumpHostId: json['jumpHostId'] as String?,
       shellIntegration: (json['shellIntegration'] as bool?) ?? true,
+      agentForwarding: (json['agentForwarding'] as bool?) ?? false,
       sftpMode: parseSftpMode(),
       sftpServerCommand: json['sftpServerCommand'] as String?,
     );
@@ -125,12 +129,13 @@ class Host {
     int? port,
     String? username,
     AuthType? authType,
-    String? keyId,
+    Object? keyId = const _Unset(),
     String? group,
     String? detectedOs,
     bool? autoRecord,
     Object? jumpHostId = const _Unset(),
     bool? shellIntegration,
+    bool? agentForwarding,
     SftpMode? sftpMode,
     Object? sftpServerCommand = const _Unset(),
   }) =>
@@ -141,7 +146,7 @@ class Host {
         port: port ?? this.port,
         username: username ?? this.username,
         authType: authType ?? this.authType,
-        keyId: keyId ?? this.keyId,
+        keyId: keyId is _Unset ? this.keyId : keyId as String?,
         group: group ?? this.group,
         tags: tags,
         createdAt: createdAt,
@@ -149,6 +154,7 @@ class Host {
         autoRecord: autoRecord ?? this.autoRecord,
         jumpHostId: jumpHostId is _Unset ? this.jumpHostId : jumpHostId as String?,
         shellIntegration: shellIntegration ?? this.shellIntegration,
+        agentForwarding: agentForwarding ?? this.agentForwarding,
         sftpMode: sftpMode ?? this.sftpMode,
         sftpServerCommand: sftpServerCommand is _Unset
             ? this.sftpServerCommand